Move HTTPClientSpy into separate file and share it between tests
