+++ Tests/Core/Isolated/Feed API/FeedsTests.swift
@@ -191,30 +191,4 @@ class FeedsTests: XCTestCase {
 
         wait(for: [expectation], timeout: 1)
     }
-
-    private class HTTPClientSpy: HTTPClient {
-        var messages: [(url: URL, completion: ((HTTPClient.Result) -> Void)?)] = []
-
-        var requestedURLs: [URL] {
-            messages.map(\.url)
-        }
-
-        func get(from url: URL, completion: ((HTTPClient.Result) -> Void)?) {
-            messages.append((url, completion))
-        }
-
-        func complete(with error: Error, at index: Int = 0) {
-            messages[index].completion?(.failure(error))
-        }
-
-        func complete(withStatusCode code: Int, data: Data, at index: Int = 0) {
-            let response = HTTPURLResponse(
-                url: messages[index].url,
-                statusCode: code,
-                httpVersion: nil,
-                headerFields: nil
-            )!
-            messages[index].completion?(.success((response, data)))
-        }
-    }
 }
